fixed smells

+++ src/main/java/com/coinsearch/service/ChainService.java
@@ -12,6 +12,7 @@ import java.util.List;
 @Service
 public class ChainService {
     private final ChainRepository chainRepository;
+    private final String ERROR_MESSAGE = "Chain does not exist with given id: ";
 
     public Chain createChain(Chain chain) {
         return chainRepository.save(chain);
@@ -20,7 +21,7 @@ public class ChainService {
     public Chain getChainById(Long chainId) {
         return chainRepository.findById(Math.toIntExact(chainId))
                 .orElseThrow(()->
-                        new EntityNotFoundException("Chain does not exist with given id: " + chainId));
+                        new EntityNotFoundException(ERROR_MESSAGE + chainId));
     }
 
 
@@ -31,7 +32,7 @@ public class ChainService {
 
     public Chain updateChain(Long chainId, Chain updatedChain) {
         Chain chain = chainRepository.findById(Math.toIntExact(chainId)).orElseThrow(
-                () -> new EntityNotFoundException("Chain does not exist with given id: " + chainId)
+                () -> new EntityNotFoundException(ERROR_MESSAGE + chainId)
         );
 
         chain.setName(updatedChain.getName());
@@ -42,9 +43,11 @@ public class ChainService {
 
 
     public void deleteChain(Long chainId) {
-        chainRepository.findById(Math.toIntExact(chainId)).orElseThrow(
-                () -> new EntityNotFoundException("Chain does not exist with given id: " + chainId)
+        Chain chain = chainRepository.findById(Math.toIntExact(chainId)).orElseThrow(
+                () -> new EntityNotFoundException(ERROR_MESSAGE + chainId)
         );
-        chainRepository.deleteById(Math.toIntExact(chainId));
+        if (chain != null) {
+            chainRepository.deleteById(Math.toIntExact(chainId));
+        }
     }
 }
