+++ src/main/java/com/coinsearch/service/PersonService.java
@@ -14,6 +14,7 @@ import java.util.List;
 public class PersonService {
 
     private PersonRepository personRepository;
+    private final String ERROR_MESSAGE = "Person does not exist with given id: ";
 
     public Person createPerson(Person person) {
         return personRepository.save(person);
@@ -23,7 +24,7 @@ public class PersonService {
     public Person getPersonById(Long personId) {
         return personRepository.findById(Math.toIntExact(personId))
                 .orElseThrow(()->
-                        new EntityNotFoundException("Person does not exist with given id: " + personId));
+                        new EntityNotFoundException(ERROR_MESSAGE + personId));
     }
 
 
@@ -34,7 +35,7 @@ public class PersonService {
 
     public Person updatePerson(Long personId, Person updatedPerson) {
         Person person = personRepository.findById(Math.toIntExact(personId)).orElseThrow(
-                () -> new EntityNotFoundException("Person does not exist with given id: " + personId)
+                () -> new EntityNotFoundException(ERROR_MESSAGE + personId)
         );
 
         person.setName(updatedPerson.getName());
@@ -45,9 +46,11 @@ public class PersonService {
 
 
     public void deletePerson(Long personId) {
-        personRepository.findById(Math.toIntExact(personId)).orElseThrow(
-                () -> new EntityNotFoundException("Person does not exist with given id: " + personId)
+        Person person = personRepository.findById(Math.toIntExact(personId)).orElseThrow(
+                () -> new EntityNotFoundException(ERROR_MESSAGE + personId)
         );
-        personRepository.deleteById(Math.toIntExact(personId));
+        if (person != null) {
+            personRepository.deleteById(Math.toIntExact(personId));
+        }
     }
 }
